back/pyproject.toml
[project]
name = "obligatorio-back"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
]
